django/requirements.txt
Django==1.11.5
django-crispy-forms==1.6.1
django-filter==1.0.4
django-reversion==2.0.10
django-tables2==1.11.0
django-taggit==0.22.1
elasticsearch==5.4.0
pytz==2017.2
urllib3==1.22
wheel==0.24.0
